Made a basic logs downloader feature
