+++ packages/hmi-frontend/src/Pages/StreamingPage/Logger/StreamLogsPage.tsx
@@ -18,29 +18,13 @@ import ky from "ky";
 import { useQuery } from "react-query";
 import { toast } from "react-toastify";
 import LogsTableCell from "./LogsTableCell";
+import downloadLogs from "@/Helper/downloadLogs";
 
 export default function StreamLogsPage() {
   const { isLoading, error, data, refetch } = useQuery<StreamLogData[], Error>(
     "logs",
     () => ky.get(`${STREAMER_SERVER}/logs`).json<StreamLogData[]>()
   );
-
-  const downloadLogs = () => {
-    const fileName = "logs";
-    const json = JSON.stringify(data, null, 2);
-    const blob = new Blob([json], { type: "application/json" });
-    const href = URL.createObjectURL(blob);
-
-    const link = document.createElement("a");
-    link.href = href;
-    link.download = fileName + ".json";
-    document.body.appendChild(link);
-    link.click();
-
-    // clean up "a" element & remove ObjectURL
-    document.body.removeChild(link);
-    URL.revokeObjectURL(href);
-  };
 
   if (error) {
     toast.error(error.message);
@@ -109,7 +93,7 @@ export default function StreamLogsPage() {
       </Box>
 
       <Button
-        onClick={downloadLogs}
+        onClick={() => downloadLogs(data, "streamer_logs")}
         fullWidth
         color="error"
         variant="contained"
