+++ packages/hmi-frontend/src/Pages/WaterfallPage/WaterfallLogsCard/WaterfallLogsCard.tsx
@@ -1,7 +1,15 @@
 import { Box, Button, Card, Paper } from "@mui/material";
 import DownloadIcon from "@mui/icons-material/Download";
+import ky from "ky";
+import { WATERFALL_BACKEND_URL } from "@/Helper/consts";
+import { useQuery } from "react-query";
+import downloadLogs from "@/Helper/downloadLogs";
 
 export default function WaterfallLogsCard() {
+  const { data } = useQuery<WaterfallObject[], Error>("logs", () =>
+    ky.get(`${WATERFALL_BACKEND_URL}/logs`).json<WaterfallObject[]>()
+  );
+
   return (
     <Card
       onClick={(e) => e.stopPropagation()}
@@ -16,12 +24,13 @@ export default function WaterfallLogsCard() {
       <Box
         sx={{
           height: "60vh",
-          minWidth: '80vw',
+          minWidth: "80vw",
           overflow: "auto",
         }}
       ></Box>
 
       <Button
+        onClick={() => downloadLogs(data, "waterfall_logs")}
         fullWidth
         color="error"
         variant="contained"
